merge plot functions and fix bugs [needs docs!]
Major plot_functions merge. No more distinction between double and normal plots. Now they are in the same function and share all functionalities. Any additional functionality only needs to be implemented in one place. Also clear two of the test files that were meant for this purpose and fixed errors with the layers.